dream home project
Our project is the 3D model of a Modern house constructed by using AutoCAD. In this project, we will try to implement all the necessary things for people. We will try to design our modern house that will be a fascinating and comfortable environment. Which includes a gymnasium, an Indoor playground for children, a swimming pool, a super shop, and so on.